Move Keystore class to its own file.
